+++ spec/lib/pact_broker/client/cli/custom_thor_spec.rb
@@ -7,6 +7,10 @@ module PactBroker::Client::CLI
   end
 
   class TestThor < CustomThor
+    def self.exit_on_failure?
+      false
+    end
+
     desc 'ARGUMENT', 'This is the description'
     def test_default(argument)
       Delegate.call(argument: argument)
@@ -94,6 +98,29 @@ module PactBroker::Client::CLI
         expect(options[:pact_broker_base_url]).to eq 'http://bar'
       end
       TestThor.start(%w{test_pact_broker_client_options})
+    end
+
+    describe "when someone copy pastes from Word and uses an em dash instead of a normal dash" do
+      before do
+        allow(TestThor).to receive(:exit_with_error_code)
+      end
+
+      it "exits with an error message" do
+        expect($stdout).to receive(:puts).with(/contains an em dash/)
+        expect(TestThor).to receive(:exit_with_error_code)
+        TestThor.start(["test_default", "\u2014\u2014bar"])
+      end
+    end
+
+    describe ".handle_help" do
+      context "when the last argument is --help or -h" do
+        it "turns it into the form that Thor expects, which is a really odd one" do
+          expect(TestThor.handle_help(["foo", "--help"])).to eq ["help", "foo"]
+          expect(TestThor.handle_help(["foo", "-h"])).to eq ["help", "foo"]
+          expect(TestThor.handle_help(["-h"])).to eq ["help"]
+          expect(TestThor.handle_help(["--help"])).to eq ["help"]
+        end
+      end
     end
 
     describe ".turn_muliple_tag_options_into_array" do